Modularise System.Random
- System.Random.Internal contains all code
- System.Random exports just the subset of the library that was present
  in v1.1, for backwards-compatibility
- System.Random.Monad exports everything present in System.Random, plus
  MonadRandom, the new monadic adapters and all MonadRandom-related
  functions

Haddocks are structured as follows:
- System.Random.Internal is hidden
- System.Random contains a notice that it exists for backwards
  compatibility, and a link to System.Random.Monad for the full
  documentation
- System.Random.Monad contains the full, structured Haddocks
