+++ random.cabal
@@ -28,7 +28,11 @@ custom-setup
         cabal-doctest >=1.0.6 && <1.1
 
 library
-    exposed-modules:  System.Random
+    exposed-modules:
+        System.Random
+        System.Random.Internal
+        System.Random.Monad
+
     c-sources:        cbits/CastFloatWord.cmm
     default-language: Haskell2010
     ghc-options:      -Wall
